# sstable,db: introduce a sstable-internal ObsoleteBit in the key kind