@sumeerbhola 2023-06-01
I've marked this as ready for review since much of the cleanup is done. My next steps:
- Add more unit tests.
- Investigate the correctness TODO in compaction_iter.go.

I'll ping when these steps are done.